requirements.txt
telethon
sqlalchemy
python-dotenv
pymysql